+++ backend/src/main/java/com/pickpick/channel/application/ChannelSubscriptionService.java
@@ -7,11 +7,11 @@ import com.pickpick.channel.domain.ChannelSubscriptionRepository;
 import com.pickpick.channel.ui.dto.ChannelOrderRequest;
 import com.pickpick.channel.ui.dto.ChannelResponse;
 import com.pickpick.channel.ui.dto.ChannelSubscriptionRequest;
-import com.pickpick.exception.ChannelNotFoundException;
-import com.pickpick.exception.MemberNotFoundException;
-import com.pickpick.exception.SubscriptionDuplicateException;
-import com.pickpick.exception.SubscriptionNotExistException;
-import com.pickpick.exception.SubscriptionOrderDuplicateException;
+import com.pickpick.exception.channel.ChannelNotFoundException;
+import com.pickpick.exception.channel.SubscriptionDuplicateException;
+import com.pickpick.exception.channel.SubscriptionNotExistException;
+import com.pickpick.exception.channel.SubscriptionOrderDuplicateException;
+import com.pickpick.exception.member.MemberNotFoundException;
 import com.pickpick.member.domain.Member;
 import com.pickpick.member.domain.MemberRepository;
 import java.util.List;
