--- backend/src/main/java/com/pickpick/exception/SlackBadRequestException.java
@@ -1,8 +1,0 @@
-package com.pickpick.exception;
-
-public class SlackBadRequestException extends BadRequestException {
-
-    public SlackBadRequestException(final String message) {
-        super(message);
-    }
-}
