+++ backend/src/main/java/com/pickpick/exception/auth/InvalidTokenException.java
@@ -1,4 +1,6 @@
-package com.pickpick.exception;
+package com.pickpick.exception.auth;
+
+import com.pickpick.exception.BadRequestException;
 
 public class InvalidTokenException extends BadRequestException {
