+++ backend/src/main/java/com/pickpick/exception/channel/SubscriptionNotExistException.java
@@ -1,4 +1,6 @@
-package com.pickpick.exception;
+package com.pickpick.exception.channel;
+
+import com.pickpick.exception.BadRequestException;
 
 public class SubscriptionNotExistException extends BadRequestException {
