+++ backend/src/main/java/com/pickpick/exception/member/MemberInvalidThumbnailUrlException.java
@@ -1,4 +1,6 @@
-package com.pickpick.exception;
+package com.pickpick.exception.member;
+
+import com.pickpick.exception.BadRequestException;
 
 public class MemberInvalidThumbnailUrlException extends BadRequestException {
