+++ backend/src/main/java/com/pickpick/exception/member/MemberInvalidUsernameException.java
@@ -1,4 +1,6 @@
-package com.pickpick.exception;
+package com.pickpick.exception.member;
+
+import com.pickpick.exception.BadRequestException;
 
 public class MemberInvalidUsernameException extends BadRequestException {
