+++ backend/src/main/java/com/pickpick/exception/message/BookmarkNotFoundException.java
@@ -1,4 +1,6 @@
-package com.pickpick.exception;
+package com.pickpick.exception.message;
+
+import com.pickpick.exception.NotFoundException;
 
 public class BookmarkNotFoundException extends NotFoundException {
