+++ backend/src/main/java/com/pickpick/exception/slackevent/SlackEventNotFoundException.java
@@ -1,4 +1,6 @@
-package com.pickpick.exception;
+package com.pickpick.exception.slackevent;
+
+import com.pickpick.exception.NotFoundException;
 
 public class SlackEventNotFoundException extends NotFoundException {
