+++ backend/src/main/java/com/pickpick/slackevent/application/member/MemberChangedService.java
@@ -1,6 +1,6 @@
 package com.pickpick.slackevent.application.member;
 
-import com.pickpick.exception.MemberNotFoundException;
+import com.pickpick.exception.member.MemberNotFoundException;
 import com.pickpick.member.domain.Member;
 import com.pickpick.member.domain.MemberRepository;
 import com.pickpick.slackevent.application.SlackEvent;
