+++ backend/src/test/java/com/pickpick/auth/support/JwtTokenProviderTest.java
@@ -3,7 +3,7 @@ package com.pickpick.auth.support;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import com.pickpick.exception.InvalidTokenException;
+import com.pickpick.exception.auth.InvalidTokenException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -32,7 +32,8 @@ class JwtTokenProviderTest {
     void validateExpiredToken() {
         // given
         long memberId = 1L;
-        JwtTokenProvider expiredTokenProvider = new JwtTokenProvider("eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIiLCJuYW1lIjoiSm9obiBEb2UiLCJpYXQiOjE1MTYyMzkwMjJ9.ih1aovtQShabQ7l0cINw4k1fagApg3qLWiB8Kt59Lno",
+        JwtTokenProvider expiredTokenProvider = new JwtTokenProvider(
+                "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIiLCJuYW1lIjoiSm9obiBEb2UiLCJpYXQiOjE1MTYyMzkwMjJ9.ih1aovtQShabQ7l0cINw4k1fagApg3qLWiB8Kt59Lno",
                 0);
         String token = expiredTokenProvider.createToken(String.valueOf(memberId));
