+++ backend/src/test/java/com/pickpick/channel/application/ChannelSubscriptionServiceTest.java
@@ -3,16 +3,15 @@ package com.pickpick.channel.application;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import com.pickpick.channel.application.ChannelSubscriptionService;
 import com.pickpick.channel.domain.Channel;
 import com.pickpick.channel.domain.ChannelRepository;
 import com.pickpick.channel.domain.ChannelSubscription;
 import com.pickpick.channel.ui.dto.ChannelOrderRequest;
 import com.pickpick.channel.ui.dto.ChannelSubscriptionRequest;
-import com.pickpick.exception.ChannelNotFoundException;
-import com.pickpick.exception.SubscriptionDuplicateException;
-import com.pickpick.exception.SubscriptionNotExistException;
-import com.pickpick.exception.SubscriptionOrderDuplicateException;
+import com.pickpick.exception.channel.ChannelNotFoundException;
+import com.pickpick.exception.channel.SubscriptionDuplicateException;
+import com.pickpick.exception.channel.SubscriptionNotExistException;
+import com.pickpick.exception.channel.SubscriptionOrderDuplicateException;
 import com.pickpick.member.domain.Member;
 import com.pickpick.member.domain.MemberRepository;
 import java.util.List;
