+++ backend/src/test/java/com/pickpick/slackevent/SlackEventTest.java
@@ -3,7 +3,7 @@ package com.pickpick.slackevent;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import com.pickpick.exception.SlackEventNotFoundException;
+import com.pickpick.exception.slackevent.SlackEventNotFoundException;
 import com.pickpick.slackevent.application.SlackEvent;
 import java.util.Map;
 import java.util.stream.Stream;
@@ -14,17 +14,6 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
 class SlackEventTest {
-
-    @DisplayName("type과 subtype에 따라 SlackEvent 탐색")
-    @ParameterizedTest
-    @MethodSource("methodSource")
-    void findSlackEventByTypeAndSubtype(final Map<String, Object> request, final SlackEvent expected) {
-        // given & when
-        SlackEvent actual = SlackEvent.of(request);
-
-        // then
-        assertThat(actual).isEqualTo(expected);
-    }
 
     private static Stream<Arguments> methodSource() {
         return Stream.of(
@@ -37,6 +26,17 @@ class SlackEventTest {
                 Arguments.of(Map.of("event", Map.of("type", "channel_deleted")), SlackEvent.CHANNEL_DELETED),
                 Arguments.of(Map.of("event", Map.of("type", "user_profile_changed")), SlackEvent.MEMBER_CHANGED)
         );
+    }
+
+    @DisplayName("type과 subtype에 따라 SlackEvent 탐색")
+    @ParameterizedTest
+    @MethodSource("methodSource")
+    void findSlackEventByTypeAndSubtype(final Map<String, Object> request, final SlackEvent expected) {
+        // given & when
+        SlackEvent actual = SlackEvent.of(request);
+
+        // then
+        assertThat(actual).isEqualTo(expected);
     }
 
     @DisplayName("존재하지 않는 SlackEvent type일 경우 예외 발생")
